+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Changed default inference device to AUTO in https://github.com/openvinotoolkit/anomalib/pull/1534
+
 ### Deprecated
 
 - Support only Python 3.10 and greater in https://github.com/openvinotoolkit/anomalib/pull/1299